+++ test/unit/rpm/header_test.rb
@@ -23,10 +23,16 @@ describe Rupert::RPM::Header do
     header.release
   end
 
-  it "maps RPM release stored in the header" do
+  it "maps RPM os name stored in the header" do
     index.expects(:get).once.with(tags[:os])
 
     header.os
+  end
+
+  it "maps RPM architecture name stored in the header" do
+    index.expects(:get).once.with(tags[:arch])
+
+    header.arch
   end
 
   it "maps RPM uncompressed size stored in the header" do